backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.117.1",
    "groq>=0.33.0",
    "langchain-groq>=1.0.0",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "passlib[bcrypt]>=1.7.4",
    "prophet>=1.1.7",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.9",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.2",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.37.0",
]
